+++ allauth/mfa/app_settings.py
@@ -63,6 +63,16 @@ class AppSettings:
         dflt = ["recovery_codes", "totp"]
         return self._setting("SUPPORTED_TYPES", dflt)
 
+    @property
+    def WEBAUTHN_ALLOW_INSECURE_ORIGIN(self):
+        return self._setting("WEBAUTHN_ALLOW_INSECURE_ORIGIN", False)
+
+    @property
+    def PASSKEY_LOGIN_ENABLED(self) -> bool:
+        return "webauthn" in self.SUPPORTED_TYPES and self._setting(
+            "PASSKEY_LOGIN_ENABLED", False
+        )
+
 
 _app_settings = AppSettings("MFA_")