+++ allauth/mfa/forms.py
@@ -1,13 +1,26 @@
+from typing import Callable
+
 from django import forms
 from django.utils.translation import gettext_lazy as _
 
 from allauth.account.adapter import get_adapter as get_account_adapter
 from allauth.account.models import EmailAddress
 from allauth.core import context, ratelimit
-from allauth.mfa import totp
+from allauth.mfa import app_settings, totp, webauthn
 from allauth.mfa.adapter import get_adapter
 from allauth.mfa.internal import flows
 from allauth.mfa.models import Authenticator
+
+
+def _check_rate_limit(user) -> Callable[[], None]:
+    key = f"mfa-auth-user-{str(user.pk)}"
+    if not ratelimit.consume(
+        context.request,
+        action="login_failed",
+        key=key,
+    ):
+        raise get_account_adapter().validation_error("too_many_login_attempts")
+    return lambda: ratelimit.clear(context.request, action="login_failed", key=key)
 
 
 class BaseAuthenticateForm(forms.Form):
@@ -23,19 +36,15 @@ class BaseAuthenticateForm(forms.Form):
         super().__init__(*args, **kwargs)
 
     def clean_code(self):
-        key = f"mfa-auth-user-{str(self.user.pk)}"
-        if not ratelimit.consume(
-            context.request,
-            action="login_failed",
-            key=key,
-        ):
-            raise get_account_adapter().validation_error("too_many_login_attempts")
-
+        clear_rl = _check_rate_limit(self.user)
         code = self.cleaned_data["code"]
-        for auth in Authenticator.objects.filter(user=self.user):
+        for auth in Authenticator.objects.filter(user=self.user).exclude(
+            # WebAuthn cannot validate manual codes.
+            type=Authenticator.Type.WEBAUTHN
+        ):
             if auth.wrap().validate_code(code):
                 self.authenticator = auth
-                ratelimit.clear(context.request, action="login_failed", key=key)
+                clear_rl()
                 return code
 
         raise get_adapter().validation_error("incorrect_code")
@@ -104,3 +113,91 @@ class GenerateRecoveryCodesForm(forms.Form):
         if not flows.recovery_codes.can_generate_recovery_codes(self.user):
             raise get_adapter().validation_error("cannot_generate_recovery_codes")
         return cleaned_data
+
+
+class AddWebAuthnForm(forms.Form):
+    name = forms.CharField(required=False)
+    if app_settings.PASSKEY_LOGIN_ENABLED:
+        passwordless = forms.BooleanField(
+            label=_("Passwordless"),
+            required=False,
+            help_text=_(
+                "Enabling passwordless operation allows you to sign in using just this key/device, but imposes additional requirements such as biometrics or PIN protection."
+            ),
+        )
+    credential = forms.JSONField(required=True, widget=forms.HiddenInput)
+
+    def __init__(self, *args, **kwargs):
+        self.user = kwargs.pop("user")
+        self.registration_data = webauthn.begin_registration(self.user)
+        initial = kwargs.setdefault("initial", {})
+        initial.setdefault(
+            "name",
+            get_adapter().generate_authenticator_name(
+                self.user, Authenticator.Type.WEBAUTHN
+            ),
+        )
+        super().__init__(*args, **kwargs)
+
+    def clean(self):
+        cleaned_data = super().clean()
+        credential = cleaned_data.get("credential")
+        if credential:
+            # Explicitly parse JSON payload -- otherwise, register_complete()
+            # crashes with some random TypeError and we don't want to do
+            # Pokemon-style exception handling.
+            webauthn.parse_registration_response(credential)
+            webauthn.complete_registration(credential)
+        return cleaned_data
+
+
+class AuthenticateWebAuthnForm(forms.Form):
+    credential = forms.JSONField(required=True, widget=forms.HiddenInput)
+
+    def __init__(self, *args, **kwargs):
+        self.user = kwargs.pop("user")
+        self.authentication_data = webauthn.begin_authentication(self.user)
+        super().__init__(*args, **kwargs)
+
+    def clean_credential(self):
+        credential = self.cleaned_data["credential"]
+        # Explicitly parse JSON payload -- otherwise, authenticate_complete()
+        # crashes with some random TypeError and we don't want to do
+        # Pokemon-style exception handling.
+        webauthn.parse_authentication_response(credential)
+        user = self.user
+        if user is None:
+            user = webauthn.extract_user_from_response(credential)
+        clear_rl = _check_rate_limit(user)
+        authenticator = webauthn.complete_authentication(user, credential)
+        clear_rl()
+        return authenticator
+
+    def save(self):
+        authenticator = self.cleaned_data["credential"]
+        authenticator.record_usage()
+
+
+class LoginWebAuthnForm(AuthenticateWebAuthnForm):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, user=None, **kwargs)
+
+
+class ReauthenticateWebAuthnForm(AuthenticateWebAuthnForm):
+    pass
+
+
+class EditWebAuthnForm(forms.Form):
+    name = forms.CharField(required=True)
+
+    def __init__(self, *args, **kwargs):
+        self.instance = kwargs.pop("instance")
+        initial = kwargs.setdefault("initial", {})
+        initial.setdefault("name", self.instance.wrap().name)
+        super().__init__(*args, **kwargs)
+
+    def save(self) -> Authenticator:
+        auth = self.instance.wrap()
+        auth.name = self.cleaned_data["name"]
+        self.instance.save()
+        return self.instance