+++ allauth/mfa/internal/flows/__init__.py
@@ -1,4 +1,9 @@
-from allauth.mfa.internal.flows import authentication, recovery_codes, totp
+from allauth.mfa.internal.flows import (
+    authentication,
+    recovery_codes,
+    totp,
+    webauthn,
+)
 
 
-__all__ = ["authentication", "recovery_codes", "totp"]
+__all__ = ["authentication", "recovery_codes", "totp", "webauthn"]