+++ allauth/mfa/stages.py
@@ -3,6 +3,8 @@ from django.urls import NoReverseMatch, reverse
 
 from allauth import app_settings as allauth_settings
 from allauth.account.stages import LoginStage
+from allauth.mfa.internal import flows
+from allauth.mfa.models import Authenticator
 from allauth.mfa.utils import is_mfa_enabled
 
 
@@ -11,7 +13,7 @@ class AuthenticateStage(LoginStage):
 
     def handle(self):
         response, cont = None, True
-        if is_mfa_enabled(self.login.user):
+        if self._should_handle(self.request):
             try:
                 response = HttpResponseRedirect(reverse("mfa_authenticate"))
             except NoReverseMatch:
@@ -21,3 +23,12 @@ class AuthenticateStage(LoginStage):
                 else:
                     raise
         return response, cont
+
+    def _should_handle(self, request) -> bool:
+        if not is_mfa_enabled(
+            self.login.user, [Authenticator.Type.TOTP, Authenticator.Type.WEBAUTHN]
+        ):
+            return False
+        if flows.authentication.did_use_passwordless_login(request):
+            return False
+        return True