+++ allauth/mfa/urls.py
@@ -9,6 +9,8 @@ urlpatterns: List[Union[URLPattern, URLResolver]] = [
     path("authenticate/", views.authenticate, name="mfa_authenticate"),
     path("reauthenticate/", views.reauthenticate, name="mfa_reauthenticate"),
 ]
+if app_settings.PASSKEY_LOGIN_ENABLED:
+    urlpatterns.append(path("login/", views.login, name="mfa_login"))
 
 if app_settings.SUPPORTED_TYPES:
     urlpatterns.append(path("", views.index, name="mfa_index"))
@@ -47,5 +49,40 @@ if "recovery_codes" in app_settings.SUPPORTED_TYPES:
                     ),
                 ]
             ),
-        ),
+        )
+    )
+
+if "webauthn" in app_settings.SUPPORTED_TYPES:
+    urlpatterns.append(
+        path(
+            "webauthn/",
+            include(
+                [
+                    path("", views.list_webauthn, name="mfa_list_webauthn"),
+                    path("add/", views.add_webauthn, name="mfa_add_webauthn"),
+                    path(
+                        "reauthenticate/",
+                        views.reauthenticate_webauthn,
+                        name="mfa_reauthenticate_webauthn",
+                    ),
+                    path(
+                        "<int:pk>/",
+                        include(
+                            [
+                                path(
+                                    "remove/",
+                                    views.remove_webauthn,
+                                    name="mfa_remove_webauthn",
+                                ),
+                                path(
+                                    "edit/",
+                                    views.edit_webauthn,
+                                    name="mfa_edit_webauthn",
+                                ),
+                            ]
+                        ),
+                    ),
+                ]
+            ),
+        )
     )